Move modal into components
