+++ x-pack/plugins/observability/public/components/slo/delete_confirmation_modal/slo_delete_confirmation_modal.stories.tsx
