+++ x-pack/plugins/observability/public/pages/slos/components/slo_list_item.tsx
@@ -21,6 +21,7 @@ import { ALL_VALUE, HistoricalSummaryResponse, SLOWithSummaryResponse } from '@k
 import type { Rule } from '@kbn/triggers-actions-ui-plugin/public';
 import { useQueryClient } from '@tanstack/react-query';
 import React, { useState } from 'react';
+import { SloDeleteConfirmationModal } from '../../../components/slo/delete_confirmation_modal/slo_delete_confirmation_modal';
 import { rulesLocatorID, sloFeatureId } from '../../../../common';
 import { SLO_BURN_RATE_RULE_TYPE_ID } from '../../../../common/constants';
 import { sloKeys } from '../../../hooks/slo/query_key_factory';
@@ -37,7 +38,6 @@ import {
   transformSloResponseToCreateSloForm,
 } from '../../slo_edit/helpers/process_slo_form_values';
 import { SloBadges } from './badges/slo_badges';
-import { SloDeleteConfirmationModal } from './slo_delete_confirmation_modal';
 import { SloSummary } from './slo_summary';
 
 export interface SloListItemProps {
